requirements.txt
# Hierarchical Financial AI System Requirements
# Core LangChain dependencies
langchain>=0.1.0
langchain-openai>=0.1.0  
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# LLM Provider APIs
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.3.0
langchain-google-genai>=1.0.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Financial data sources
yfinance>=0.2.20
ccxt>=4.0.0

# HTTP client and utilities
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Database (for RAG)
pymongo>=4.5.0

# Additional utilities
asyncio-mqtt>=0.16.0  # For potential real-time features
tiktoken>=0.5.0       # For token counting
